jellyfish-cli/com.ngc.seaside.jellyfish.utilities/src/test/java/com/ngc/seaside/jellyfish/utilities/file/FileUtilitiesTest.java
/**
 * UNCLASSIFIED
 *
 * Copyright 2020 Northrop Grumman Systems Corporation
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of
 * this software and associated documentation files (the "Software"), to deal in
 * the Software without restriction, including without limitation the rights to use,
 * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
 * Software, and to permit persons to whom the Software is furnished to do so,
 * subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
 * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
 * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.ngc.seaside.jellyfish.utilities.file;

import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.TemporaryFolder;

import java.io.File;
import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.ArrayList;
import java.util.List;

import static junit.framework.TestCase.assertTrue;

/**
 *
 */
public class FileUtilitiesTest {

   @Rule
   public TemporaryFolder testFolder = new TemporaryFolder();


   @Test
   public void doesAddLineToFile() throws IOException, FileUtilitiesException {
      File settingsFile = testFolder.newFile("settings.gradle");
      Path settingsPath = Paths.get(settingsFile.getAbsolutePath());

      List<String> lines = new ArrayList<>();
      lines.add("line1");
      lines.add("line2");

      FileUtilities.addLinesToFile(settingsPath, lines);

      List<String> newLines = Files.readAllLines(settingsPath);

      assertTrue(newLines.contains("line1"));
      assertTrue(newLines.contains("line2"));
   }

}
